data_collection/requirements.txt
tldextract
pyenchant
compound-word-splitter
numpy
pandas
dnspython
pyasn
netaddr
flask
networkx
google-cloud-storage
cython
fastparquet
gcsfs
google-cloud-pubsub
cloudstorage